ZEPHYR-28943: Implemented MasterToSlaveFileCallable to deliver callable code to slave machine and execute code there
