+++ src/main/java/com/thed/service/BaseService.java
@@ -1,9 +1,11 @@
 package com.thed.service;
 
+import java.io.Serializable;
+
 /**
  * Created by prashant on 25/6/19.
  */
-public interface BaseService {
+public interface BaseService extends Serializable {
 
     ZephyrRestService getZephyrRestService();
